fix: resolve constants refactor issues
- Fix password validation error message for SDK tests
- Add missing HTTP status constants
- Fix service name format for health tests
- Correct Strava rate limits to match production values
- Remove sample data from src/ for production hygiene
- Fix protocol module re-exports to avoid conflicts

All tests passing except 4 MCP multitenant tests with HTTP 500 errors

+++ src/api_keys.rs
@@ -57,7 +57,7 @@ impl ApiKeyTier {
     #[must_use]
     pub const fn default_trial_days(&self) -> Option<i64> {
         match self {
-            Self::Trial => Some(TRIAL_PERIOD_DAYS as i64), // Trial period
+            Self::Trial => Some(TRIAL_PERIOD_DAYS), // Trial period
             _ => None,
         }
     }
